Python/Django/biblioteca/requirements.txt
Django>=5.2
django-widget-tweaks
djangorestframework>=3.14.0